requirements.txt
cffi==1.17.1
colorama==0.4.6
cryptography==43.0.3
numpy==2.1.3
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
six==1.16.0
tzdata==2024.2
